upd8

+++ run.py
@@ -1,5 +1,5 @@
 from pathlib import Path
-
+import pandas as pd
 from dataset import create_df
 
 dataset_path = Path("./resources")
@@ -8,8 +8,11 @@ image_root_path = dataset_path / "common-mobile-web-app-icons"
 
 def main():
     df_path = dataset_path / "data.csv"
-    df = create_df(root_path=image_root_path)
-    df.to_csv(df_path, index=False)
+    if df_path.exists():
+        df = pd.read_csv(df_path)
+    else:
+        df = create_df(root_path=image_root_path)
+        df.to_csv(df_path, index=False)
 
 
 if __name__ == '__main__':
